Added tools folder, tools Makefile and bootstrap.cpp for importing blocks from a bootstrap.dat file.
usage: bootstrap <blockchain directory> <bootstrap.dat file>

bootstrap.cpp will check block chain directory and init the chain with the genesis block if it has not already been initiated. Then it will begin importing all blocks from the bootstrap.dat file. First block in the bootstrap.dat file is checked to ensure it matches the block hash of genesis_block().

Sometimes the loop stalls waiting for a future to be fulfilled (line 183 & 184), not sure why, but this needs investigating.