Track previous stable release, not next version.
Note that previous v2.0 still has a bad tag string. `git describe`
works, however.

From http://nvie.com/posts/a-successful-git-branching-model/

    It is exactly at the start of a release branch that the upcoming
    release gets assigned a version number—not any earlier. Up until
    that moment, the develop branch reflected changes for the “next
    release”, but it is unclear whether that “next release” will
    eventually become 0.3 or 1.0, until the release branch is started.
    That decision is made on the start of the release branch and is
    carried out by the project’s rules on version number bumping.

This is so if a critical bug is discovered for which the fix is still
compatible with the previous major release, the fix can be applied
to the master branch without bumping the major version, but bumping the
Z in X.Y.Z instead.