Allow user-defined mnemonic wordlists
The previous enum-based mnemonic-creation approach limited the user to just
the wordlists provided in libbitcoin, which puts wallet authors in a bind.

For instance, suppose Bitcoin usage explodes in Greece,
and the industry quickly responds with a Greek wordlist.
The enum-based approach gives the wallet author no choice but to
patch and upgrade libbitcoin if they want Greek, which is scary from
a release-management point of view. With this approach,
the wallet author can provide their own Greek wordlist right away,
and then upgrade libbitcoin when the next regular release comes out.

This also simplifies the validation logic, removing ambiguity problems.
A mnemonic is valid if it passes validation in any language,
removing the need to pre-detect which language the mnemonic belongs to.