Fixes electrum mnemonic validation by requiring the initial entropy.
Validation is now as follows: Provided a word list and a starting
entropy value, it is possible to create a new mnemonic, which should
match the provided word list.

There are a couple electrum test cases that do not provide the
original entropy, so we don't validate those.  Many more electrum test
cases do provide the entropy, so those are now properly validated.